Move rest of Board in Game
