+++ src/test/java/com/javachess/logic/PawnTests.java
@@ -25,7 +25,7 @@ class PawnTests {
   @Test
   @DisplayName("canMoveTo should return true if a white pawn moves one step up and no piece is at that position")
   public void canMoveToWhiteOneStepUp() {
-    Game g = Game.of(1, Player.of("white", Color.WHITE), Player.of("black", Color.BLACK), Board.of());
+    Game g = Game.of(1, Player.of("white", Color.WHITE), Player.of("black", Color.BLACK));
     Piece pawn = Game.getPieceAt("d", "2", g).get();
     assertEquals(true, Pawn.canMoveTo("d", "3", Game.getPieces(g), pawn));
   }
@@ -33,7 +33,7 @@ class PawnTests {
   @Test
   @DisplayName("canMoveTo should return false if a white pawn moves three steps up")
   public void canMoveToWhiteThreeStepsUp() {
-    Game g = Game.of(1, Player.of("white", Color.WHITE), Player.of("black", Color.BLACK), Board.of());
+    Game g = Game.of(1, Player.of("white", Color.WHITE), Player.of("black", Color.BLACK));
     Piece pawn = Game.getPieceAt("d", "2", g).get();
     assertEquals(false, Pawn.canMoveTo("d", "5", Game.getPieces(g), pawn));
   }
@@ -41,7 +41,7 @@ class PawnTests {
   @Test
   @DisplayName("canMoveTo should return false if a white pawn moves two steps up and it is not in its initial position anymore")
   public void canMoveToWhiteTwoStepsUpAfterInitialPosition() {
-    Game g = Game.of(1, Player.of("white", Color.WHITE), Player.of("black", Color.BLACK), Board.of());
+    Game g = Game.of(1, Player.of("white", Color.WHITE), Player.of("black", Color.BLACK));
     g = Game.doMove(Move.of(Position.of("d", "2"), Position.of("d", "3")), g);
     Piece pawn = Game.getPieceAt("d", "3", g).get();
     assertEquals(false, Pawn.canMoveTo("d", "5", Game.getPieces(g), pawn));
@@ -50,7 +50,7 @@ class PawnTests {
   @Test
   @DisplayName("canMoveTo should return false if a piece is present at destination")
   public void canMoveToPieceAtDestination() {
-    Game g = Game.of(1, Player.of("white", Color.WHITE), Player.of("black", Color.BLACK), Board.of());
+    Game g = Game.of(1, Player.of("white", Color.WHITE), Player.of("black", Color.BLACK));
     g = Game.doMove(Move.of(Position.of("d", "7"), Position.of("d", "4")), g);
     Piece pawn = Game.getPieceAt("d", "2", g).get();
     assertEquals(false, Pawn.canMoveTo("d", "4", Game.getPieces(g), pawn));
@@ -59,7 +59,7 @@ class PawnTests {
   @Test
   @DisplayName("canMoveTo should return true if a white piece moves one step in diagonal and a black piece is there")
   public void canMoveToPieceDiagonal() {
-    Game g = Game.of(1, Player.of("white", Color.WHITE), Player.of("black", Color.BLACK), Board.of());
+    Game g = Game.of(1, Player.of("white", Color.WHITE), Player.of("black", Color.BLACK));
     g = Game.doMove(Move.of(Position.of("e", "7"), Position.of("e", "3")), g);
     g = Game.doMove(Move.of(Position.of("c", "7"), Position.of("c", "3")), g);
     Piece pawn = Game.getPieceAt("d", "2", g).get();
@@ -70,7 +70,7 @@ class PawnTests {
   @Test
   @DisplayName("canMoveTo should return false if a white piece moves one step in diagonal and a white piece is there")
   public void canMoveToPieceDiagonalSameColor() {
-    Game g = Game.of(1, Player.of("white", Color.WHITE), Player.of("black", Color.BLACK), Board.of());
+    Game g = Game.of(1, Player.of("white", Color.WHITE), Player.of("black", Color.BLACK));
     g = Game.doMove(Move.of(Position.of("e", "2"), Position.of("e", "3")), g);
     Piece pawn = Game.getPieceAt("d", "2", g).get();
     assertEquals(false, Pawn.canMoveTo("e", "3", Game.getPieces(g), pawn));
@@ -79,7 +79,7 @@ class PawnTests {
   @Test
   @DisplayName("canMoveTo should return false if a white piece moves one step in diagonal and no piece is there")
   public void canMoveToPieceDiagonalEmptySquare() {
-    Game g = Game.of(1, Player.of("white", Color.WHITE), Player.of("black", Color.BLACK), Board.of());
+    Game g = Game.of(1, Player.of("white", Color.WHITE), Player.of("black", Color.BLACK));
     Piece pawn = Game.getPieceAt("d", "2", g).get();
     assertEquals(false, Pawn.canMoveTo("e", "3", Game.getPieces(g), pawn));
   }
@@ -87,7 +87,7 @@ class PawnTests {
   @Test
   @DisplayName("canMoveTo should return true if a black piece moves one step down")
   public void canMoveToPieceBlackDown() {
-    Game g = Game.of(1, Player.of("white", Color.WHITE), Player.of("black", Color.BLACK), Board.of());
+    Game g = Game.of(1, Player.of("white", Color.WHITE), Player.of("black", Color.BLACK));
     Piece pawn = Game.getPieceAt("d", "7", g).get();
     assertEquals(true, Pawn.canMoveTo("d", "6", Game.getPieces(g), pawn));
   }
@@ -95,7 +95,7 @@ class PawnTests {
   @Test
   @DisplayName("canMoveTo should return false if a black pawn moves two steps down and it is not in its initial position anymore")
   public void canMoveToBlackTwoStepsDownAfterInitialPosition() {
-    Game g = Game.of(1, Player.of("white", Color.WHITE), Player.of("black", Color.BLACK), Board.of());
+    Game g = Game.of(1, Player.of("white", Color.WHITE), Player.of("black", Color.BLACK));
     g = Game.doMove(Move.of(Position.of("d", "7"), Position.of("d", "6")), g);
     Piece pawn = Game.getPieceAt("d", "6", g).get();
     assertEquals(false, Pawn.canMoveTo("d", "4", Game.getPieces(g), pawn));
@@ -104,7 +104,7 @@ class PawnTests {
   @Test
   @DisplayName("canMoveTo should return false if a black piece moves more than two steps down")
   public void canMoveToPieceBlack3Down() {
-    Game g = Game.of(1, Player.of("white", Color.WHITE), Player.of("black", Color.BLACK), Board.of());
+    Game g = Game.of(1, Player.of("white", Color.WHITE), Player.of("black", Color.BLACK));
     Piece pawn = Game.getPieceAt("d", "7", g).get();
     assertEquals(false, Pawn.canMoveTo("d", "4", Game.getPieces(g), pawn));
   }
